pyproject.toml
[project]
name = "fluencyfeatureannotator"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "RusCucumber", email = "w.d.ryu.m@gmail.com" }
]
dependencies = [
    "flet>=0.24.1",
    "google-cloud-storage==1.42.2",
    "google-cloud-speech==2.7.0",
    "praat-textgrids==1.3.1",
    "syllables==1.0.3",
    "pydub==0.25.1",
    "stanza==1.2.3",
    "werkzeug==2.0.3",
    "rev-ai==2.15.0",
    "tqdm==4.64.1",
    "inflect==5.3.0",
    "matplotlib==3.4.3",
    "spacy>=3.8.2",
    "numpy>=2.0.2",
    "sentencepiece>=0.2.0",
    "pip>=24.2",
    "graphviz>=0.20.3",
    "torch>=2.4.1",
    "pandas>=2.2.3",
    "praat-parselmouth>=0.4.4",
    "webrtcvad>=2.0.10",
    "spacy-transformers>=1.3.5",
    "transformers==4.20.1",
    "tokenizers==0.11.6",
    "jupyter>=1.1.1",
    "unidecode>=1.3.8",
    "cmudict>=1.0.31",
    "g2p-en>=2.1.0",
    "torchaudio>=2.4.1",
    "soundfile==0.12.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.6.4",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/fluencyfeatureannotator"]

[tool.ruff]
target-version = "py311"
line-length = 119
exclude = [".mypy_cache"]
 
[tool.ruff.lint]
select = ["E", "W", "F", "I", "C90"]
ignore = ["E203"]
